py_backend/requirements.txt
fastapi
uvicorn
openai
requests
sse-starlette
python-dotenv
authlib